everything works fine

+++ NTierApp.ASPMVC/Models/Company/AddCompanyToUserViewModel.cs
@@ -7,7 +7,7 @@ namespace NTierApp.ASPMVC.Models.Company
 {
     public class AddCompanyToUserViewModel
     {
-        public int UserId { get; set; }
-        public int CompanyId { get; set; }
+        public long UserId { get; set; }
+        public long CompanyId { get; set; }
     }
 }
